requirements.txt
transformers
torch
sentencepiece
gradio
langdetect
pytesseract
Pillow
